+++ oshi-core/src/main/java/oshi/util/platform/unix/solaris/KstatUtil.java
@@ -56,12 +56,8 @@ public final class KstatUtil {
 
     private static final Logger LOG = LoggerFactory.getLogger(KstatUtil.class);
 
-    private static final LibKstat KS = LibKstat.INSTANCE;
-
-    // Opens the kstat chain. Automatically closed on exit.
     // Only one thread may access the chain at any time, so we wrap this object in
     // the KstatChain class which locks the class until closed.
-    private static final KstatCtl KC = KS.kstat_open();
     public static final ReentrantLock CHAIN = new ReentrantLock();
 
     private KstatUtil() {
@@ -72,15 +68,17 @@ public final class KstatUtil {
      * one thread may actively use this object at any time.
      * <p>
      * The chain is created once using the {@link KstatUtil#openChain} method.
-     * Instantiating this object (after locking) updates the chain and is the
-     * equivalent of calling {@link LibKstat#kstat_open}. The control object should
-     * be closed with {@link #close}, the equivalent of calling
-     * {@link LibKstat#kstat_close}
+     * Instantiating this object opens the chain calling
+     * {@link LibKstat#kstat_open}. The control object should be closed with
+     * {@link #close}, the equivalent of calling {@link LibKstat#kstat_close}
      */
     public static final class KstatChain implements AutoCloseable {
 
+        private KstatCtl kstatCtl = null;
+
         private KstatChain() {
-            update();
+            CHAIN.lock();
+            kstatCtl = LibKstat.INSTANCE.kstat_open();
         }
 
         /**
@@ -96,9 +94,9 @@ public final class KstatUtil {
          *            The kstat from which to retrieve data
          * @return {@code true} if successful; {@code false} otherwise
          */
-        public static boolean read(Kstat ksp) {
+        public boolean read(Kstat ksp) {
             int retry = 0;
-            while (0 > KS.kstat_read(KC, ksp, null)) {
+            while (0 > LibKstat.INSTANCE.kstat_read(kstatCtl, ksp, null)) {
                 if (LibKstat.EAGAIN != Native.getLastError() || 5 <= ++retry) {
                     if (LOG.isDebugEnabled()) {
                         LOG.debug("Failed to read kstat {}:{}:{}",
@@ -128,8 +126,8 @@ public final class KstatUtil {
          * @return The first match of the requested Kstat structure if found, or
          *         {@code null}
          */
-        public static Kstat lookup(String module, int instance, String name) {
-            return KS.kstat_lookup(KC, module, instance, name);
+        public Kstat lookup(String module, int instance, String name) {
+            return LibKstat.INSTANCE.kstat_lookup(kstatCtl, module, instance, name);
         }
 
         /**
@@ -149,9 +147,10 @@ public final class KstatUtil {
          * @return All matches of the requested Kstat structure if found, or an empty
          *         list otherwise
          */
-        public static List<Kstat> lookupAll(String module, int instance, String name) {
+        public List<Kstat> lookupAll(String module, int instance, String name) {
             List<Kstat> kstats = new ArrayList<>();
-            for (Kstat ksp = KS.kstat_lookup(KC, module, instance, name); ksp != null; ksp = ksp.next()) {
+            for (Kstat ksp = LibKstat.INSTANCE.kstat_lookup(kstatCtl, module, instance, name); ksp != null; ksp = ksp
+                    .next()) {
                 if ((module == null || module.equals(Native.toString(ksp.ks_module, StandardCharsets.US_ASCII)))
                         && (instance < 0 || instance == ksp.ks_instance)
                         && (name == null || name.equals(Native.toString(ksp.ks_name, StandardCharsets.US_ASCII)))) {
@@ -162,24 +161,11 @@ public final class KstatUtil {
         }
 
         /**
-         * Convenience method for {@link LibKstat#kstat_chain_update}. Brings this kstat
-         * header chain in sync with that of the kernel.
-         * <p>
-         * This function compares the kernel's current kstat chain ID(KCID), which is
-         * incremented every time the kstat chain changes, to this object's KCID.
-         *
-         * @return the new KCID if the kstat chain has changed, 0 if it hasn't, or -1 on
-         *         failure.
-         */
-        public static int update() {
-            return KS.kstat_chain_update(KC);
-        }
-
-        /**
-         * Release the lock on the chain.
+         * Close the chain and release the lock on the chain.
          */
         @Override
         public void close() {
+            LibKstat.INSTANCE.kstat_close(kstatCtl);
             CHAIN.unlock();
         }
     }
@@ -191,7 +177,6 @@ public final class KstatUtil {
      *         are done with it with {@link KstatChain#close()}.
      */
     public static KstatChain openChain() {
-        CHAIN.lock();
         return new KstatChain();
     }
 
@@ -213,7 +198,7 @@ public final class KstatUtil {
         if (ksp.ks_type != LibKstat.KSTAT_TYPE_NAMED && ksp.ks_type != LibKstat.KSTAT_TYPE_TIMER) {
             throw new IllegalArgumentException("Not a kstat_named or kstat_timer kstat.");
         }
-        Pointer p = KS.kstat_data_lookup(ksp, name);
+        Pointer p = LibKstat.INSTANCE.kstat_data_lookup(ksp, name);
         if (p == null) {
             LOG.debug("Failed to lookup kstat value for key {}", name);
             return "";
@@ -257,7 +242,7 @@ public final class KstatUtil {
         if (ksp.ks_type != LibKstat.KSTAT_TYPE_NAMED && ksp.ks_type != LibKstat.KSTAT_TYPE_TIMER) {
             throw new IllegalArgumentException("Not a kstat_named or kstat_timer kstat.");
         }
-        Pointer p = KS.kstat_data_lookup(ksp, name);
+        Pointer p = LibKstat.INSTANCE.kstat_data_lookup(ksp, name);
         if (p == null) {
             if (LOG.isDebugEnabled()) {
                 LOG.debug("Failed lo lookup kstat value on {}:{}:{} for key {}",